remove_text
To delete multiple lines that contain specific sentences in a text file using Python, you can modify the code to include a list of sentences to look for. Here's an example of how you could modify the previous code:

This code reads in the text file using Python's built-in open() function and sets a list of sentences to look for. It then loops over each line in the file, checking if the line contains any of the sentences using the any() function. If the line contains any of the sentences, it skips it and moves on to the next line. If the line does not contain any of the sentences, it appends it to a new string. Finally, it writes the modified text to a new file using Python's built-in open() function. Note that this code will remove all lines that contain any of the sentences, so make sure to double-check the output before using it.